make DryadJournalConcept a subclass of a more generic DryadOrganizationConcept
